requirements.txt
Django==1.10
django-markdownx==1.6.2
Markdown==2.6.6
Pillow==3.3.0
psycopg2==2.6.2
whitenoise==3.2.1
